Implement better searching for bolt executable
Previously, `bolt` needed to be in the PATH for this task to work. This
commit updates the task to search two expected locations for the bolt
executable and call them directly if found, before defaulting back to
relying on PATH should the executable not be found in any of the
expected locations.